flattening padding bugfix and clear missing in align

+++ src/flatnormals.h
@@ -6,10 +6,10 @@
 #include <vector>
 
 void flattenRadialNormals(int w, int h, std::vector<float> &normals, double binSize = 20);
-void flattenFourierNormals(int w, int h, std::vector<float> &normals, int padding = 0.2, double sigma = 20, bool exponential = true);
+void flattenFourierNormals(int w, int h, std::vector<float> &normals, float padding = 0.2, double sigma = 20, bool exponential = true);
 
 void flattenRadialHeights(int w, int h, std::vector<float> &heights, double binSize = 20);
-void flattenFourierHeights(int w, int h, std::vector<float> &heights, int padding = 0.2, double sigma = 20);
+void flattenFourierHeights(int w, int h, std::vector<float> &heights, float padding = 0.2, double sigma = 20);
 
 class NormalsImage {
 public:
